Improve testability of statistics and remove WcdWikibaseModel to simplify the hierarchy and setup the wikibase attribute when we need it.
